requirements.txt
python-telegram-bot[async]==21.0.1
pymongo==4.6.0
python-dotenv==1.0.0
Pillow==10.1.0
networkx==3.2.1
matplotlib==3.8.2
APScheduler==3.10.4
aiohttp==3.9.1
requests==2.31.0